Separate Redis driver (in progress...)
